Ignore unknown syscalls on aarch64
The auditd rules as indicated in the CIS rules, don't work
on AL2023 Gravitron (aarch64) instances. The unavailable syscalls:
 - unlink
 - rename
 - creat
 - open
 - chown
 - lchown
 - chmod

For full consistency, I've updated the checks to support the previously
used auditd rule, and also added a second rule with the unavailable
syscall removed.

The stime syscall is also not available, but in cis-dil-benchmark-4.1.5
this syscall is already excluded from the line.

Signed-off-by: Ivo van Doorn <ivovandoorn@samotics.com>