Extend checks for the filesystem configuration
all checks for the filesystem configurations validate
that the kernel module is not loaded, however, the cis
description indicates
 => If this filesystem type is not needed, disable it.

This would make it fair if we add a check that validates
that the given module is not mentioned in /etc/fstab.

This resolves failures on AL2023 where /boot/efi is loaded
using the `vfat` filesystem type.

Signed-off-by: Ivo van Doorn <ivovandoorn@samotics.com>